requirements.txt
absl-py==2.1.0
aif360==0.6.1
aiocoap==0.4.7
altgraph==0.17.3
amqp==5.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
av==12.0.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comet-ml==3.47.3
comm==0.2.2
configobj==5.0.9
contourpy==1.0.7
cycler==0.11.0
DateTime==5.5
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dlib==19.24.4
dulwich==0.22.6
entrypoints==0.4
et-xmlfile==1.1.0
everett==3.1.0
exceptiongroup==1.2.2
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
fsspec==2024.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.34.0
imbalanced-learn==0.12.4
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.6.17
keras==3.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.4.2
lazy_loader==0.3
libclang==18.1.1
logging==0.4.9.6
macholib==1.16.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pillow_heif==0.21.0
platformdirs==4.3.6
polarTransform==2.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.4.0
Pygments==2.18.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
python-box==6.1.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.19.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
surrealdb==0.3.2
svgpathtools==1.6.0
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2024.2.12
tinycss2==1.3.0
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchinfo==1.8.0
torchvision==0.15.2
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
ucimlrepo==0.0.6
ultralytics==8.3.37
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==2.2.3
vine==5.1.0
virtualenv==20.21.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wrapt==1.16.0
wurlitzer==3.1.1
xlrd==2.0.1
zope.interface==6.3
